Documented classes and methods (except getters).

+++ include/birdgame/distance_util.h
@@ -7,7 +7,16 @@
 
 namespace birdgame {
 
-float GetEuclideanDistance(float x1, float y1, float x2, float y2);
+// This function calculates and returns the Euclidean distance
+// between points with coordinates of the passed floats.
+
+// @param x1 the x-coordinate of the first point
+// @param y1 the y-coordinate of the first point
+// @param x2 the x-coordinate of the second point
+// @param y2 the y-coordinate of the second point
+
+// @return the calculated Euclidean distance between the points
+size_t GetEuclideanDistance(float x1, float y1, float x2, float y2);
 
 }  // namespace birdgame
