+++ src/distance_util.cc
@@ -5,9 +5,10 @@
 namespace birdgame {
 
 // Code inspired by Mineopoly
-float GetEuclideanDistance(float x1, float y1, float x2, float y2) {
+size_t GetEuclideanDistance(float x1, float y1, float x2, float y2) {
+
     // The 2 is used to square (x1 - x2) and (y1 - y2), as per the distance formula.
-    return sqrt(pow((x1 - x2), 2) + pow((y1 - y2), 2));
+    return (size_t) sqrt(pow((x1 - x2), 2) + pow((y1 - y2), 2));
 }
 
 }  // namespace birdgame
